requirements.txt
Pillow
natsort
tqdm
matplotlib
numpy
segmentation_models_pytorch
albumentations
tensorboard

marshmallow_dataclass
hydra-core